Fix functionality bug on edit command after find
Currently, there is an error in the code where the edit
command executes twice.

This results in the wrong candidate records changing if
edit command comes immediately after the find command.

Let's fix the error by ensuring that the edit command
only executes once by adding a boolean check in the MainWindow.